Confirmation message when deleting data

+++ AndersonCRMWeb/Scripts/Angular/Controller/Employee.Controller.js
@@ -113,19 +113,23 @@
         }
 
         function Delete(employeeId) {
-            EmployeeService.Delete(employeeId)
-                .then(function (response) {
-                    Read();
-                })
-                .catch(function (data, status) {
-                    new PNotify({
-                        title: status,
-                        text: data,
-                        type: 'error',
-                        hide: true,
-                        addclass: "stack-bottomright"
+            var conf = window.confirm("Are you sure you want to delete?");
+            if (conf == true) {
+                EmployeeService.Delete(employeeId)
+                    .then(function (response) {
+                        Read();
+                    })
+                    .catch(function (data, status) {
+                        new PNotify({
+                            title: status,
+                            text: data,
+                            type: 'error',
+                            hide: true,
+                            addclass: "stack-bottomright"
+                        });
                     });
-                });
+            }
+            else { return; false}
         }
 
     }
